Refactor superblock code, update data structure
The kernel panic was a result of not allocating a root node (which still
hasn't been implemented.)  I refactored the super block code to separate
the module loading code from the actual super block methods.  I also
updated the data structures, because of the difference between what is
stored on disk and what needs to be in memory.  For example, there is
now one block data structure; all of the filesystem metadata resides in
data structures that keep pointers to blocks.