Implemented checkpoint block # calculations
I implemented some helper functions to set the number of checkpoint
blocks in the on-disk superblock; some of them will also be used at
mount time, so I put them in their own util module.  I also split the
block data structure into on-disk & in-memory data structures.  In order
to determine the number of segments, you must know the size of the block
device being formatted; for actual block devices, this is accomplished
with an ioctl, while for loop devices (files,) it must be done using a
stat call.