Clean up mkfs code, RPi2 compatible build
Minor refactor of the mkfs-wlfs helpers.  Also fixed the build for the
Raspberry Pi 2 (armv7-a.)  Kernel code for the RPi2 is soft float for
portability reasons I imagine, but the RPi2 is hard float (armhf,) and
the native gcc seemingly will not produce code that doesn't use the VFP
registers.  This meant that util.o couldn't be linked with mkfs-wlfs.o,
since the former doesn't use VFP registers when compiled for a kernel
module.  After discovering that cross compiling for armel is a giant
pain on Raspbian and Arch Linux ARM, I found that the simple solution is
to simply compile util.c twice.