Fix Debian setup & build
Installing gtest is slightly more complicated than installing a package
on Debian; the important bit is that you have to compile the provided
source into shared libraries.  Also, pthreads was causing linking
errors, so I disabled it in the build.  mkfs-wlfs is no longer as strict
about block size, since there doesn't seem to be an easy/portable way to
figure out the physical sector size of the block device that a file
resides on (from C.)