Update segmap block calculation
There's no point in keeping per-segment bitmaps; they're only computed &
used when cleaning.  The filesystem just needs to be able to easily find
the next free segment during normal operations.  I added macros to make
sure that there is at least one imap/segmap block, and changed the way
that the number of segments & checkpoint blocks are determined.  The
number of segments is calculated initially and used to produce an upper
bound on the number of checkpoint blocks necessary; once this is known,
the number of segments available is decremented by the number of
checkpoint segments.  Before, either the number of segments or
checkpoint blocks was actually being used without being initialized,
which led to some undefined behavior on the Raspberry Pi.  This solution
wastes some disk space, but we're talking about some extra bits in the
last checkpoint (segmap) block